parameterized fnprovider
