+++ codegen/smithy-go-codegen/src/main/java/software/amazon/smithy/go/codegen/SmithyGoDependency.java
@@ -43,6 +43,7 @@ public final class SmithyGoDependency {
     public static final GoDependency XML = stdlib("encoding/xml");
     public static final GoDependency SYNC = stdlib("sync");
     public static final GoDependency PATH = stdlib("path");
+    public static final GoDependency LOG = stdlib("log");
 
     public static final GoDependency SMITHY = smithy(null, "smithy");
     public static final GoDependency SMITHY_TRANSPORT = smithy("transport", "smithytransport");
@@ -62,7 +63,7 @@ public final class SmithyGoDependency {
     public static final GoDependency SMITHY_DOCUMENT_JSON = smithy("document/json", "smithydocumentjson");
     public static final GoDependency SMITHY_SYNC = smithy("sync", "smithysync");
     public static final GoDependency SMITHY_AUTH_BEARER = smithy("auth/bearer");
-    public static final GoDependency SMITHY_ENDPOINT_RULESFN = smithy("endpoint/rulesfn");
+    public static final GoDependency SMITHY_ENDPOINT_RULESFN = smithy("endpoints/private/rulesfn");
 
     public static final GoDependency GO_CMP = goCmp("cmp");
     public static final GoDependency GO_CMP_OPTIONS = goCmp("cmp/cmpopts");
