+++ codegen/smithy-go-codegen/src/main/java/software/amazon/smithy/go/codegen/endpoints/FnProvider.java
@@ -1,0 +1,7 @@
+package software.amazon.smithy.go.codegen.endpoints;
+
+import software.amazon.smithy.codegen.core.Symbol;
+
+public interface FnProvider {
+    Symbol fnFor(String name);
+}
